refactor: simplify error responses
