+++ backend/src/main/java/com/github/mdeluise/plantit/exception/ControllerExceptionHandler.java
@@ -1,55 +1,45 @@
 package com.github.mdeluise.plantit.exception;
 
-import java.util.Date;
-
+import com.github.mdeluise.plantit.exception.error.ErrorCode;
+import com.github.mdeluise.plantit.exception.error.ErrorMessage;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.context.request.WebRequest;
 
 @ControllerAdvice
 @ResponseBody
 public class ControllerExceptionHandler {
 
     @ExceptionHandler(ResourceNotFoundException.class)
-    public ResponseEntity<ErrorMessage> entityNotFoundException(
-        ResourceNotFoundException ex,
-        WebRequest request) {
-        ErrorMessage message = new ErrorMessage(
-                HttpStatus.NOT_FOUND.value(),
-                new Date(),
-                ex.getMessage(),
-                request.getDescription(false),
-                ex.getCause() != null ? ex.getCause().getMessage() : null
+    public ResponseEntity<ErrorMessage> entityNotFoundException(ResourceNotFoundException ex) {
+        final ErrorMessage message = new ErrorMessage(
+            HttpStatus.NOT_FOUND.value(),
+            ErrorCode.RESOURCE_NOT_FOUND,
+            ex.getMessage()
         );
         return new ResponseEntity<>(message, HttpStatus.NOT_FOUND);
     }
 
+
     @ExceptionHandler(UnauthorizedException.class)
-    public ResponseEntity<ErrorMessage> unauthorizedException(
-        UnauthorizedException ex,
-        WebRequest request) {
-        ErrorMessage message = new ErrorMessage(
+    public ResponseEntity<ErrorMessage> unauthorizedException(UnauthorizedException ex) {
+        final ErrorMessage message = new ErrorMessage(
             HttpStatus.UNAUTHORIZED.value(),
-            new Date(),
-            ex.getMessage(),
-            request.getDescription(false),
-            ex.getCause() != null ? ex.getCause().getMessage() : null
+            ErrorCode.USER_UNAUTHORIZED,
+            ex.getMessage()
         );
         return new ResponseEntity<>(message, HttpStatus.UNAUTHORIZED);
     }
 
 
     @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorMessage> globalExceptionHandler(Exception ex, WebRequest request) {
-        ErrorMessage message = new ErrorMessage(
-                HttpStatus.INTERNAL_SERVER_ERROR.value(),
-                new Date(),
-                ex.getMessage(),
-                request.getDescription(false),
-                ex.getCause() != null ? ex.getCause().getMessage() : null
+    public ResponseEntity<ErrorMessage> globalExceptionHandler(Exception ex) {
+        final ErrorMessage message = new ErrorMessage(
+            HttpStatus.INTERNAL_SERVER_ERROR.value(),
+            ErrorCode.INTERNAL_SERVER_ERROR,
+            ex.getMessage()
         );
         return new ResponseEntity<>(message, HttpStatus.INTERNAL_SERVER_ERROR);
     }
