--- backend/src/main/java/com/github/mdeluise/plantit/exception/ErrorMessage.java
@@ -1,6 +1,0 @@
-package com.github.mdeluise.plantit.exception;
-
-import java.util.Date;
-
-public record ErrorMessage(int statusCode, Date timestamp, String message, String description, String cause) {
-}
