+++ backend/src/main/java/com/github/mdeluise/plantit/exception/error/ErrorCode.java
@@ -1,0 +1,8 @@
+package com.github.mdeluise.plantit.exception.error;
+
+public enum ErrorCode {
+    RESOURCE_NOT_FOUND,
+    USER_UNAUTHORIZED,
+    INTERNAL_SERVER_ERROR,
+    AUTHENTICATION_ERROR,
+}
