+++ backend/src/main/java/com/github/mdeluise/plantit/exception/error/ErrorMessage.java
@@ -1,0 +1,4 @@
+package com.github.mdeluise.plantit.exception.error;
+
+public record ErrorMessage(int statusCode, ErrorCode errorCode, String message) {
+}
